progress

+++ src/TodoList.sol
@@ -9,29 +9,29 @@ contract Todolist {
     }
 
     struct Todo {
-        uint id;
+        uint256 id;
         string content;
         Status status;
     }
 
-    event TodoAdded(uint _id, string _content);
+    event TodoAdded(uint256 _id, string _content);
 
-    mapping(uint => Todo) public todos;
-    uint[] public todoIds;
+    mapping(uint256 => Todo) public todos;
+    uint256[] public todoIds;
 
     function addTodos(string memory _content) public {
-        uint newId = todoIds.length;
+        uint256 newId = todoIds.length;
         todos[newId] = Todo(newId, _content, Status.Pending);
         todoIds.push(newId);
         emit TodoAdded(newId, _content);
     }
 
-    function updateStatus(uint _id, Status _status) public {
+    function updateStatus(uint256 _id, Status _status) public {
         require(_id < todoIds.length, "Todo does not exist");
         todos[_id].status = _status;
     }
 
-    function getTodo(uint _id) public view returns (Todo memory) {
+    function getTodo(uint256 _id) public view returns (Todo memory) {
         require(_id < todoIds.length, "Todo does not exist");
         return todos[_id];
     }
